+++ VDA-Name-Registry/ChangeLog.md
@@ -1,4 +1,9 @@
+2023-03-24 (v0.3.1)
+-------------------
+- Moved `EnumerableSet` & `StringLib` into the `VDA-Common` contract.
+
 2023-03-15 (v0.3.0)
 -------------------
 - Multi-sign support added
 - Solidity version upgraded to '0.8.18'
+- Hardhat configuration updated